+++ CHANGES.rst
@@ -7,6 +7,8 @@ Version 1.0.0.dev0 (in development)
 
 - Added
 
+  * a :class:`~msl.network.client.Link` can create an exclusive or shared lock
+    with a :class:`~msl.network.service.Service`
   * the :attr:`~msl.network.network.Device.loop_thread_id` property for a
     :class:`~msl.network.service.Service` and a :class:`~msl.network.client.Client`
   * the :meth:`~msl.network.service.Service.emit_notification_threadsafe` method